requirements.txt
# Flask and Extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug>=3.0.0

# Database
SQLAlchemy>=2.0.0
psycopg2-binary==2.9.9

# Environment
python-dotenv==1.0.0

# AI/OpenAI
openai==1.12.0

# Image Processing
Pillow==10.2.0
opencv-python-headless==4.9.0.80

# Data Science
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0

# Visualization
networkx==3.2.1
matplotlib==3.8.2
plotly==5.18.0

# Security
cryptography==42.0.0

# Utilities
python-multipart==0.0.6

# Production Server
gunicorn==21.2.0
